refactor(runner): isolate terminal job logging
